Force GWT 2.0.4 for now, start to remove some JSNI code.
It turns out that GWT 2.1 allows display code to be initialized later on.
Our canvas hack assumes that the canvas is visible at a certain
point, and with the new version it no longer is.

GWT 2.2 looks like it will have a native canvas implementation and
we can switch to that.  In the meantime, these changes have us using
native GWT code for keypress handling, which is a good thing either
way.